\name{scDataviz-package}

\title{scDataviz: single cell dataviz and downstream analyses}

\description{In the 'single cell' World, which includes flow cytometry, mass cytometry, single-cell RNA-seq (scRNA-seq), and others, there is a need to improve data visualisation of results and to bring analysis capabilities to researchers even from non-technical backgrounds who have some experience in coding. *scDataviz* [@scDataviz] attempts to fit into this space, while also catering for advanced users. Due to the way that it's designed, *scDataviz* also has a 'plug and play' feel, whereby the base storage unit, which is based on *SingleCellExperiment* [@Lun], immediately lends flexibility and compatibility with studies that go beyond *scDataviz*. Additionally, the graphics are generated via the *ggplot* [@Wickham] engine, which means that users can 'add on' features to these to their pleasing.}

\docType{package}
